pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hfl"
version = "0.12.2"
description = "Run HuggingFace models locally like Ollama"
license = {file = "LICENSE"}
requires-python = ">=3.10"
dependencies = [
    # Using ~= (compatible release) for stability: allows patch updates only
    "typer>=0.15.0,<1.0",
    "rich~=13.0",
    # Pinned to <1.0 because transformers 5.x switched its huggingface_hub
    # contract (>=1.5.0) and is incompatible with the rest of HFL's
    # surface, which relies on the 0.x API. See ``test_dependency_pins.py``
    # for the static guard.
    "huggingface-hub>=0.27.0,<1.0",
    "fastapi~=0.135.0",
    "starlette>=0.49.1",
    "python-multipart>=0.0.22",
    "uvicorn[standard]~=0.34.0",
    "pydantic~=2.12.0",
    "httpx~=0.28.0",
    "pyyaml~=6.0",
]

[project.optional-dependencies]
# llama-cpp-python requires compilation - install separately if needed.
# Pinned to >=0.3.20 because earlier versions bundle a llama.cpp build
# that doesn't recognise the ``gemma4`` architecture and rejects Gemma 4
# GGUFs at load time with "unknown model architecture".
llama = ["llama-cpp-python>=0.3.20"]
transformers = [
    # Pinned to the 4.x line: transformers 5.x requires huggingface_hub
    # >= 1.5.0 (different API) which is incompatible with the rest of
    # HFL. The 4.x line targets huggingface_hub 0.x which we use in our
    # main dependency. See ``test_dependency_pins.py`` for the static
    # guard that prevents this pin from being widened by accident.
    "transformers>=4.47.0,<5.0",
    "torch>=2.5.0",
    "accelerate>=1.2.0",
    "sentencepiece>=0.2.0",
]
vllm = ["vllm>=0.6.0"]
convert = ["gguf>=0.10.0"]

# TTS support
tts = [
    # Same 4.x pin as the [transformers] extra — see comment there.
    "transformers>=4.47.0,<5.0",
    "torch>=2.5.0",
    "torchaudio>=2.2.0",
    "soundfile>=0.12.1",
]
coqui = [
    "coqui-tts>=0.24.0",
    "torch>=2.5.0",
    "torchaudio>=2.2.0",
    "soundfile>=0.12.1",
]
# Audio playback (optional)
audio = [
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
]

tray = [
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
]

# Model Context Protocol client/server (Phase 9+)
mcp = ["mcp>=1.0.0"]

# Apple Silicon native backend (Phase 13+ — V2 row 14)
mlx = [
    "mlx-lm>=0.18.0; sys_platform == 'darwin' and platform_machine == 'arm64'",
]

# Speech-to-text (Whisper) — Phase 16 / V2 row 16. Prefers
# faster-whisper (CPU+GPU); openai-whisper stays as a fallback.
stt = [
    "faster-whisper>=1.0.0",
]

# Image generation (SDXL / FLUX) — Phase 16 / V2 row 17. Heavy: a
# fresh install drags in ~2 GB of wheels. Opt-in only.
imagegen = [
    "diffusers>=0.30.0",
    "torch>=2.5.0",
    "safetensors>=0.4.0",
    "Pillow>=10.0.0",
    "accelerate>=1.2.0",
]

# OpenTelemetry tracing — Phase 17 / V2 row 24.
otel = [
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-exporter-otlp-proto-http>=1.27.0",
]

all = ["hfl[llama,transformers,vllm,convert,tts,coqui,audio,tray,mcp,mlx,stt,imagegen]"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24.0", "ruff>=0.8.0", "pytest-cov>=4.0", "mypy>=1.10.0", "numpy>=1.24.0"]
build = ["pyinstaller>=6.0"]

[project.scripts]
hfl = "hfl.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/hfl"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "--cov=hfl --cov-report=term-missing --cov-report=html --cov-report=xml"
markers = [
    "acceptance: tool-calling acceptance suite (spec §6, T1–T7)",
    "slow: tests that take noticeable wall-clock time (sleeps, concurrency)",
    "integration: tests that exercise filesystem, subprocesses, or system probes",
]

[tool.coverage.run]
source = ["src/hfl"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    # TTS engines require external deps (torch, transformers, coqui-tts)
    "src/hfl/engine/bark_engine.py",
    "src/hfl/engine/coqui_engine.py",
    # CLI is tested via integration tests
    "src/hfl/cli/main.py",
    "src/hfl/cli/commands/*",
    # Tray requires GUI (pystray) - tested with mocks
    "src/hfl/tray/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    # Don't require coverage for abstract methods
    "@abstractmethod",
]
fail_under = 75
show_missing = true

[tool.coverage.html]
directory = "htmlcov"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = false  # Gradually enable
check_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = [
    "hfl.api.*",
    "hfl.cli.commands.*",
]
disallow_untyped_defs = true
